Authentication, JWT, Auth middleware, (compleate crud functionality with authentication).

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from "../models/userModel.js"
 
 // @desc    GET goals
 // @route   GET /api/goals
 // @access  Private
 export const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -19,6 +20,7 @@ export const setGoal = asyncHandler(async (req, res) => {
       }
       const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
       })
     
     res.status(200).json(goal)
@@ -34,6 +36,21 @@ export const updateGoal = asyncHandler(async (req, res) => {
       res.status(400)
       throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+    
+    //check for user
+    if(!user) {
+      res.status(401)
+      throw new Error('User not found')
+    }
+     
+    // Make sure the
+    if(goal.user.toString() !== user.id) {
+       res.status(401)
+       throw new Error('User not authorised')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
       })
@@ -50,6 +67,21 @@ export const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error('Goal not found');
     }
+
+    const user = await User.findById(req.user.id)
+    
+    //check for user
+    if(!user) {
+      res.status(401)
+      throw new Error('User not found')
+    }
+     
+    // Make sure the
+    if(goal.user.toString() !== user.id) {
+       res.status(401)
+       throw new Error('User not authorised')
+    }
+    
     await goal.deleteOne();
 
     res.status(200).json({ id: req.params.id });
